Updates made.

+++ Miscellaneous/ModularExponentiation.cpp
@@ -2,7 +2,7 @@
 // For arbitrary x,y and m
 //In general m = 1e9+7
 
-long long modExpo (int x, int y, int m) {
+long long modExpo (long long x, long long y, long long m) {
     if (y == 0)
         return 1;
